toggle calendar view feature
